requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
datasets==1.5.0
dill==0.3.3
filelock==3.0.12
flatbuffers==1.12
fsspec==0.9.0
future==0.18.2
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
huggingface-hub==0.0.8
idna==2.10
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
matplotlib==3.4.3
multidict==5.1.0
multiprocess==0.70.11.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
Pillow==8.3.2
protobuf==3.15.8
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-box==5.3.0
python-dateutil==2.8.1
pytorch-lightning==1.1.8
pytz==2021.1
PyYAML==5.3.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.44
scikit-learn==0.24.1
scipy==1.6.2
seqeval==1.2.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
tokenizers==0.10.2
torch==1.8.1
tqdm==4.49.0
transformers==4.5.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
xxhash==2.0.0
yarl==1.6.3
